requirements.txt
azure-iot-device
sqlalchemy
colorlog
Flask